pyproject.toml
[tool.poetry]
name = "studies-schedule-sync"
version = "0.1.0"
description = ""
authors = ["Michał Oręziak <micorix.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
requests = "^2.28.1"
black = "^22.10.0"
flake8 = "^5.0.4"
google-api-python-client = "^2.64.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pyre-check = "^0.9.16"
black = "^22.10.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
